+++ gulp/tasks/html.js
@@ -8,7 +8,34 @@ import htmlmin from 'gulp-htmlmin'
 import htmlbeautify from 'gulp-html-beautify'
 import path from 'path'
 import fs from 'fs'
-import { getPaths, handleErrors } from '../libs/utils'
+import {
+	getPaths,
+	handleErrors
+} from '../libs/utils'
+import r from 'nunjucks/src/runtime'
+
+function SafeString(val) {
+	if(typeof val !== 'string') {
+		return val
+	}
+
+	this.val = val
+	this.length = val.length
+}
+
+SafeString.prototype = Object.create(String.prototype, {
+	length: {
+		writable: true,
+		configurable: true,
+		value: 0
+	}
+})
+SafeString.prototype.valueOf = function () {
+	return this.val
+}
+SafeString.prototype.toString = function () {
+	return this.val
+}
 
 
 export function htmlTask() {
@@ -23,20 +50,36 @@ export function htmlTask() {
 		const dataPath = path.resolve(process.env.PWD, PATH_CONFIG.src, PATH_CONFIG.html.src, TASK_CONFIG.html.dataFile)
 		return JSON.parse(fs.readFileSync(dataPath, 'utf8'))
 	}
-	
-	const manageEnv = function(environment) {
-	
+
+	const manageEnv = function (environment) {
+
+		/*
+			array methods (nunjuck filters)
+			"things": [{a: 1}, {a: 2}, {a: 3}]
+			{{ things|has('a', 2) }}
+			{% for thing in things %}
+				<p>{{ thing.a }}</p>
+			{% endfor %}
+			<p>2</p>
+		*/
 		environment.addFilter('has', (input, prop, value) => {
 			return input.filter((entry) => entry[prop] === value)
 		})
-	
+
+		/*
+			the opposite of has
+		*/
 		environment.addFilter('not', (input, prop, value) => {
 			return input.filter((entry) => entry[prop] !== value)
 		})
-	
+
+		/*
+			Randomise array
+			{{ things|rand }}
+		*/
 		environment.addFilter('rand', (input) => {
 			let i = input.length
-			while (i) {
+			while(i) {
 				let j = Math.floor(Math.random() * i)
 				let t = input[(i -= 1)]
 				input[i] = input[j]
@@ -44,9 +87,32 @@ export function htmlTask() {
 			}
 			return input
 		})
-	
+
+		/*
+			Limit the number of items returned from array
+			{{ things|limit(2) }}
+		*/
 		environment.addFilter('limit', (input, count) => input.slice(0, count))
 
+		/*
+			string methods (nunjuck filters)
+			"thing": "Hello\nHugh"
+			<p>{{ thing|nbr2tag('p') }}</p>
+			<p>Hello</p><p>Hugh</p>
+		*/
+		environment.addFilter('nbr2tag', (input, tag) => {
+			return r.markSafe(input.replace(/\n/i, `</${tag}><${tag}>`))
+		})
+
+		/*
+			string methods (nunjuck filters)
+			"thing": "Hello\nHugh"
+			{{ thing|nbr2br }}
+			Hello<br />Hugh
+		*/
+		environment.addFilter('nbr2br', (input) => {
+			return r.markSafe(input.replace(/\n/i, '<br />'))
+		})
 	}
 
 
@@ -55,6 +121,7 @@ export function htmlTask() {
 		.on('error', handleErrors)
 		.pipe(render({
 			path: [path.resolve(process.env.PWD, PATH_CONFIG.src, PATH_CONFIG.html.src)],
+			ext: '.njk',
 			envOptions: {
 				watch: false
 			},